move Investors component to own folder
