+++ src/App.tsx
@@ -3,34 +3,9 @@ import styled from 'styled-components'
 import {Market, TermsOfService} from './pages'
 import {Nav} from './components'
 import Hero from './components/Hero/Hero'
+import Investors from './components/Investors/Investors'
 import './App.scss'
 
-const Wrapper = styled.div`
-  padding-top: 90px;
-  padding-bottom: 60px;
-`
-
-const MainContainer = styled.div`
-  display: flex;
-  max-width: 1100px;
-  margin: auto;
-`
-
-const Breakline = styled.div`
-  width: 100%;
-  border-bottom: 1px solid black;
-  height: 1px;
-`
-
-const Investors = () => {
-  return (
-    <Wrapper>
-      <MainContainer>
-        <Breakline />
-      </MainContainer>
-    </Wrapper>
-  )
-}
 const App: React.FC = () => {
   return (
     <div className="App">
